oteljava - fix compilation issue
